Refactor transaction into a folder with kernel
