+++ dpc/src/testnet1/transaction/transaction.rs
@@ -43,7 +43,6 @@ use std::{
     PartialEq(bound = "C: Testnet1Components"),
     Eq(bound = "C: Testnet1Components")
 )]
-// TODO (howardwu): Remove the public visibility here
 pub struct Transaction<C: Testnet1Components> {
     /// The network this transaction is included in
     pub network: Network,
