+++ dpc/src/testnet2/program/program.rs
@@ -18,11 +18,10 @@ use crate::testnet2::Testnet2Components;
 use snarkvm_algorithms::traits::{CommitmentScheme, CRH};
 use snarkvm_fields::{ConstraintFieldError, ToConstraintField};
 
-/// Program verification key and proof
-/// Represented as bytes to be generic for any Program SNARK
+/// Program verifying key and proof, represented as bytes to be generic for any program SNARK.
 #[derive(Clone)]
 pub struct PrivateProgramInput {
-    pub verification_key: Vec<u8>,
+    pub verifying_key: Vec<u8>,
     pub proof: Vec<u8>,
 }
