+++ parameters/examples/testnet2_outer_snark.rs
@@ -74,7 +74,7 @@ pub fn setup<C: Testnet2Components>() -> Result<(Vec<u8>, Vec<u8>), DPCError> {
         rng,
     )?;
     let private_program_input = PrivateProgramInput {
-        verification_key: to_bytes![noop_program_snark_parameters.verifying_key]?,
+        verifying_key: to_bytes![noop_program_snark_parameters.verifying_key]?,
         proof: to_bytes![program_snark_proof]?,
     };
